Solve arithmetic overflow and logic operator warnings

+++ src/OpenLoco/Utility/Stream.hpp
@@ -7,7 +7,7 @@ namespace OpenLoco::Utility
     template<typename T1, typename T2, typename T3>
     std::basic_istream<T1, T2>& readData(std::basic_istream<T1, T2>& stream, T3* dst, size_t count)
     {
-        return stream.read((char*)dst, count * sizeof(T3));
+        return stream.read((char*)dst, static_cast<uint64_t>(count) * sizeof(T3));
     }
 
     template<typename T1, typename T2, typename T3>
